+++ CHANGELOG.md
@@ -47,6 +47,8 @@ Test/interim releases are not documented here.
   - https://github.com/StateSmith/StateSmith/issues/521
 - Simulator - distinguish available transition edges.
   - https://github.com/StateSmith/StateSmith/issues/522
+- Simulator - show full state description on ALT+CLICK.
+  - https://github.com/StateSmith/StateSmith/issues/523
 - Simulator - verbose mode to show enter behavior.
   - https://github.com/StateSmith/StateSmith/issues/515
 - Simulator - verbose mode to show exit behavior.